Error handling overhaul
The method of error handling is now much more verbose. Any function that
has the ability to return an error and is possible to be envoked by the
user has a trait managed by the error_handler that holds a handle()
function. The handle() function will give a more verbose and elegent
error message that does not rely on Rust's built in expect() function