Package parsing up-and-running
Using an output.log file that was created as a result of piping the
output from reqwesting pydigger, pip_frontend can now search it for
packages and parse the output (minus the reqwest part atm. The website
was very unstable the last time I used it, so I used the output file
instead).