Added viper search
viper search is now a viable command. It will first request pip_frontend
to search for all packages matching with the name you gave it.
pip_frontend will then query pydigger and parse the output into a list
of returned packages and their details that will be printed nicely by
the search command